rename package pkg/storage to uplink/storage
