+++ lib/uplink/object.go
@@ -9,9 +9,9 @@ import (
 	"time"
 
 	"storj.io/storj/internal/readcloser"
-	"storj.io/storj/pkg/storage/streams"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/uplink/metainfo/kvmetainfo"
+	"storj.io/storj/uplink/storage/streams"
 	"storj.io/storj/uplink/stream"
 )
