+++ pkg/audit/cursor.go
@@ -14,10 +14,10 @@ import (
 	"github.com/zeebo/errs"
 
 	"storj.io/storj/pkg/pb"
-	"storj.io/storj/pkg/storage/meta"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/satellite/metainfo"
 	"storj.io/storj/uplink/eestream"
+	"storj.io/storj/uplink/storage/meta"
 )
 
 // Stripe keeps track of a stripe's index and its parent segment
