+++ pkg/datarepair/repairer/repairer.go
@@ -17,13 +17,13 @@ import (
 	"storj.io/storj/pkg/identity"
 	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
-	ecclient "storj.io/storj/pkg/storage/ec"
-	"storj.io/storj/pkg/storage/segments"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/pkg/transport"
 	"storj.io/storj/satellite/metainfo"
 	"storj.io/storj/satellite/orders"
 	"storj.io/storj/storage"
+	"storj.io/storj/uplink/ecclient"
+	"storj.io/storj/uplink/storage/segments"
 )
 
 // Error is a standard error class for this package.
