+++ satellite/metainfo/metainfo.go
@@ -24,13 +24,13 @@ import (
 	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/signing"
-	"storj.io/storj/pkg/storage/meta"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/satellite/attribution"
 	"storj.io/storj/satellite/console"
 	"storj.io/storj/satellite/orders"
 	"storj.io/storj/storage"
 	"storj.io/storj/uplink/eestream"
+	"storj.io/storj/uplink/storage/meta"
 )
 
 const (
