+++ uplink/ecclient/client_planet_test.go
@@ -22,10 +22,10 @@ import (
 	"storj.io/storj/internal/testrand"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/signing"
-	ecclient "storj.io/storj/pkg/storage/ec"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/satellite"
 	"storj.io/storj/storagenode"
+	"storj.io/storj/uplink/ecclient"
 	"storj.io/storj/uplink/eestream"
 )
