+++ uplink/metainfo/kvmetainfo/objects.go
@@ -13,12 +13,12 @@ import (
 	"storj.io/storj/pkg/encryption"
 	"storj.io/storj/pkg/paths"
 	"storj.io/storj/pkg/pb"
-	"storj.io/storj/pkg/storage/meta"
-	"storj.io/storj/pkg/storage/objects"
-	"storj.io/storj/pkg/storage/segments"
-	"storj.io/storj/pkg/storage/streams"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/storage"
+	"storj.io/storj/uplink/storage/meta"
+	"storj.io/storj/uplink/storage/objects"
+	"storj.io/storj/uplink/storage/segments"
+	"storj.io/storj/uplink/storage/streams"
 )
 
 // DefaultRS default values for RedundancyScheme
