+++ uplink/storage/segments/common.go
