+++ uplink/storage/segments/mock_store.go
