+++ uplink/storage/segments/peek.go
