+++ uplink/storage/streams/eof.go
