+++ uplink/storage/streams/shim.go
@@ -10,8 +10,8 @@ import (
 
 	"storj.io/storj/pkg/encryption"
 	"storj.io/storj/pkg/ranger"
-	"storj.io/storj/pkg/storage/segments"
 	"storj.io/storj/pkg/storj"
+	"storj.io/storj/uplink/storage/segments"
 )
 
 // Store interface methods for streams to satisfy to be a store
