+++ uplink/storage/streams/size.go
